+++ src/app/chunks/interactive/chunk-checkbox-list/chunk-checkbox-list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, Input, AfterViewInit } from '@angular/core';
 import { CheckboxList } from './CheckboxList';
 import { IntersectionObserverService } from '../../../core/intersection-observer/intersection-observer.service';
 import { LessonData } from '../../../course-viewer/lesson.model';
+import { ProgressService } from 'src/app/core/progress.service';
+
 
 @Component({
   selector: 'app-chunk-checkbox-list',
@@ -12,12 +14,17 @@ export class ChunkCheckboxListComponent implements OnInit, AfterViewInit {
   @Input() attributes: CheckboxList;
   @Input() id: string;
   @Input() parentLesson: LessonData;
-  sampleItems: string[];
+  selectedItems: string[] = [];
+
+  // Si el Chunk ha sido contestado hace un toggle de todas las opciones para marcarlas
+  toggleItems = false;
 
   constructor(
-    private intersectionObserverService: IntersectionObserverService
+    private intersectionObserverService: IntersectionObserverService,
+    private progressStore: ProgressService
   ) {
     this.attributes = new CheckboxList();
+    this.selectedItems = [];
   }
 
   ngOnInit() {
@@ -26,7 +33,44 @@ export class ChunkCheckboxListComponent implements OnInit, AfterViewInit {
     if (this.attributes.backgroundColor === undefined) { this.attributes.backgroundColor = '#ffffff'; }
   }
   ngAfterViewInit() {
+    this.loadAnswers();
     this.attributes.statementData = this.attributes.items[0];
     this.intersectionObserverService.createObserver(this.id, this.attributes, this.parentLesson);
+  }
+
+  /**
+   * Carga las respuestas de Firebase después de iniciar la vista
+   * si no existe crea una por defecto para análisis
+   */
+  loadAnswers() {
+    this.progressStore.checkAnswered(this.id)
+    .subscribe(
+      progress => {
+        if (progress && progress.answer !== '') {
+          this.selectedItems = progress.answer;
+          this.toggleItems = true;
+        } else {
+          this.progressStore.setAnswer(this.id, '');
+        }
+      }
+    );
+  }
+
+  checkAllSelected(item: string) {
+    // Añade o elimina el item del array de seleccionados porque se activa en cada toggle del checkbox
+    if (this.selectedItems.findIndex(selectedItem => selectedItem === item) === -1) {
+      this.selectedItems.push(item);
+    } else {
+      this.selectedItems = this.selectedItems.filter(selectedItem => selectedItem !== item);
+    }
+
+    // Si todos los items han sido seleccionados marca el chunk como completado
+    if (this.selectedItems.length === this.attributes.items.length) {
+      this.progressStore.setAnswer(this.id, this.selectedItems);
+    } else {
+      this.progressStore.setAnswer(this.id, '');
+    }
+
+    this.progressStore.answerInteractiveChunk(true);
   }
 }